clean up datasets/utils.py by moving functions to their own files
