+++ covid_moonshot_ml/data/openeye.py
@@ -1,11 +1,15 @@
 from openeye import oechem
 
 
-def load_openeye_pdb(pdb_fn):
+def load_openeye_pdb(pdb_fn, alt_loc=False):
     ifs = oechem.oemolistream()
+    ifs_flavor = oechem.OEIFlavor_PDB_Default | oechem.OEIFlavor_PDB_DATA
+    ## Add option for keeping track of alternat locations in PDB file
+    if alt_loc:
+        ifs_flavor |= oechem.OEIFlavor_PDB_ALTLOC
     ifs.SetFlavor(
         oechem.OEFormat_PDB,
-        oechem.OEIFlavor_PDB_Default | oechem.OEIFlavor_PDB_DATA,
+        ifs_flavor,
     )
     ifs.open(pdb_fn)
     in_mol = oechem.OEGraphMol()
@@ -29,7 +33,37 @@ def load_openeye_sdf(sdf_fn):
     return coords_mol
 
 
-def split_openeye_mol(complex_mol: oechem.OEMolBase):
+def save_openeye_pdb(mol, pdb_fn):
+    ofs = oechem.oemolostream()
+    ofs.SetFlavor(oechem.OEFormat_PDB, oechem.OEOFlavor_PDB_Default)
+    ofs.open(pdb_fn)
+    oechem.OEWriteMolecule(ofs, mol)
+    ofs.close()
+
+
+def save_openeye_sdf(mol, sdf_fn):
+    ofs = oechem.oemolostream()
+    ofs.SetFlavor(oechem.OEFormat_SDF, oechem.OEOFlavor_SDF_Default)
+    ofs.open(sdf_fn)
+    oechem.OEWriteMolecule(ofs, mol)
+    ofs.close()
+
+
+def split_openeye_mol(complex_mol, lig_chain="A"):
+    """
+    Split an OpenEye-loaded molecule into protein, ligand, etc.
+
+    Parameters
+    ----------
+    complex_mol : oechem.OEMolBase
+        Complex molecule to split.
+    lig_chain : str, default="A"
+        Which copy of the ligand to keep. Pass None to keep all ligand atoms.
+
+    Returns
+    -------
+    """
+
     ## Test splitting
     lig_mol = oechem.OEGraphMol()
     prot_mol = oechem.OEGraphMol()
@@ -41,19 +75,52 @@ def split_openeye_mol(complex_mol: oechem.OEMolBase):
     opts = oechem.OESplitMolComplexOptions()
     opts.SetSplitCovalent(True)
     opts.SetSplitCovalentCofactors(True)
-    print(
-        oechem.OESplitMolComplex(
-            lig_mol, prot_mol, water_mol, oth_mol, complex_mol
+
+    ## Select protein as all protein atoms in chain A or chain B
+    prot_only = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_Protein
+    )
+    a_chain = oechem.OERoleMolComplexFilterFactory(
+        oechem.OEMolComplexChainRoleFactory("A")
+    )
+    b_chain = oechem.OERoleMolComplexFilterFactory(
+        oechem.OEMolComplexChainRoleFactory("B")
+    )
+    a_or_b_chain = oechem.OEOrRoleSet(a_chain, b_chain)
+    opts.SetProteinFilter(oechem.OEAndRoleSet(prot_only, a_or_b_chain))
+
+    ## Select ligand as all residues with resn LIG
+    lig_only = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_Ligand
+    )
+    if lig_chain is None:
+        opts.SetLigandFilter(lig_only)
+    else:
+        lig_chain = oechem.OERoleMolComplexFilterFactory(
+            oechem.OEMolComplexChainRoleFactory(lig_chain)
         )
+        opts.SetLigandFilter(oechem.OEAndRoleSet(lig_only, lig_chain))
+
+    ## Set water filter (keep all waters in A, B, or W chains)
+    ##  (is this sufficient? are there other common water chain ids?)
+    wat_only = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_Water
+    )
+    w_chain = oechem.OERoleMolComplexFilterFactory(
+        oechem.OEMolComplexChainRoleFactory("W")
+    )
+    all_wat_chains = oechem.OEOrRoleSet(a_or_b_chain, w_chain)
+    opts.SetWaterFilter(oechem.OEAndRoleSet(wat_only, all_wat_chains))
+
+    oechem.OESplitMolComplex(
+        lig_mol,
+        prot_mol,
+        water_mol,
+        oth_mol,
+        complex_mol,
+        opts,
     )
 
-    print(
-        complex_mol.NumAtoms(),
-        lig_mol.NumAtoms(),
-        prot_mol.NumAtoms(),
-        water_mol.NumAtoms(),
-        oth_mol.NumAtoms(),
-    )
     return {
         "complex": complex_mol,
         "lig": lig_mol,
@@ -61,6 +128,10 @@ def split_openeye_mol(complex_mol: oechem.OEMolBase):
         "water": water_mol,
         "other": oth_mol,
     }
+
+
+def get_ligand_rmsd_openeye(ref: oechem.OEMolBase, mobile: oechem.OEMolBase):
+    return oechem.OERMSD(ref, mobile)
 
 
 def get_ligand_rmsd_from_pdb_and_sdf(
