+++ covid_moonshot_ml/modeling.py
@@ -1,10 +1,6 @@
 from openeye import oechem, oedocking, oespruce
 
-from .datasets.utils import (
-    load_openeye_pdb,
-    load_openeye_sdf,
-    split_openeye_mol,
-)
+from .data.openeye import load_openeye_pdb, load_openeye_sdf, split_openeye_mol
 
 
 def du_to_complex(du):
